requirements.txt
Authlib==1.2.0
Flask==2.2.2
google_api_python_client==2.73.0
Pillow==9.4.0
protobuf==4.21.12
Werkzeug==2.2.2
